Restructure loop (1 genome per cycle)
Now we only evaluate the fitness of one genome per cycle,
as opposed to perorming a whole generation in that time.

This is necessary in order to allow the framework to produce the
segmented image and any other perception we may use in fitness eval.

I considered segmenting the image manually, but the function that does
so in bhuman is in asmjit, likely for a reason, and anyway simulating
the other perceptions manually for all genomes in a population is
outright impossible.

Now the calibration procedure is a tiny little state machine,
that during each update step either performs a bunch of elaboration
(like spawning a new generation, for example) or computes the fitness
of one genome (and moves on to the next state once all the population
has ben evaluated).

This compiles, but hasn't been tested yet.