Ensure: bad ColorCal => BallPerceptn istantly lost
Make sure that the perception of the ball depends entirely on the
quality of the color calibration: if the segmented image is bad,
the ball must not be recognizable.

This condition is FALSE in normal operation of the bhuman framework
because the previous ball position is always added as a potential ball
spot, bypassing recognition in the segmented image.
Therefore, the ball continues to be recognized even with a bad color
calibration as long as it stays still or doesn't move too fast.

Thus, a new module BallSpotsWithoutPreviousGuesswas implemented,
which behaves exactly like BallSpotsProvider except it does not add
the previous ball position as a ball spot,
thus making it necessary to continuously rely on the segmented image.
In this way, while the fitness of a bad genome is evaluated
the ball will NOT be visible.

The model was "implemented" in such a way that any future changes to
BallSpotsProvider automatically reflect to the new module.
The way I finally managed to satisfy this requirement within the
limitations of the bhuman framework... may be described as questionable.
See the added files for a detailed rationale of my choices.

The original BallSpotsProvider.cpp was edited slightly to achieve
this result, but the functionality of the original module
is completely unaltered.

The new module is only applied to the Lower thread
since only that camera is used for color calibration,
and is only activated in the color calibration scenario.

Minor changes:
moved robot in the calibration scene,
calibration auto-start temporarily disabled.