Add calibration saving feature
Added a final "End" state when the calibration procedure reaches
its conclusion.

Here, (one of) the best genome produced by the search is set as the
final FieldColors and saved to the configuration files permanently.

Simulation is OK.
Real test still pending, but expected behavior is that the calbration
is saved in the robot's files (so it can be reloaded even if
bhuman is reset) but not automatically to the computer from which the
simrobot command was issued. Which is why the crresponding save command
is automatically suggested in the console.