Add polynomial scheduling for parameters
Implemented the option to have some parameters change gradually
as generations pass.
These parameters are: crossover and mutation probabilities, in order to
allow the algorithm to focus on exploration first and fine-tuning later;
and selection tournament size, in order to make the selective pressure
increase over time.

Each parameter is controllable independently via cfg parameters.
The scheduling law is always polynomial, but the exponent can be chosen
freely, so that should guarantee sufficient freedom.