Finish calibration phase 2
Written phase 2 algorithm in FieldColorsCalibrator.
- New cfg parameters for this phase have been defined.
- A fully new set of CalibrationStates has been made.
- The functions that only need slight modifications are shared with
  phase 1 and have been given an extra argument to distinguish the
  phases.
- The functions that needed a larger restructuring have been made anew.
- There is the possibility to choose whether to fine-tune only the field
  threshold (the only relevant parameters) or allow everything to
  change.

New fitness for phase 2 in FieldColorsGenome: encourage to see much
green in the lower half and little green in the upper half of the image,
and at the same time don't let the fine-tuned solution stray too much
from that found in phase 1.
Also, new constructor for Genome to make one from a FieldColors object.

IMPORTANT: It seems I've been addressing the column-major
image as though it was row-major this entire time.
The issue has been corrected (now I use an overloaded
operator that internally turns to column-major).
The fix seems to have broken nothing else,
but I'll test this in real too.

Card slightly improved.