Join behaviors
Created a new scenario for the full calibration.
And a new card.

Notable: copied the CameraControlEngine code into
AutomaticCameraCalibrator2 in order to allow the latter to process
head motion requests from the cards. This may not necessarily be best
practice, but it works for the project, and I personally have no better
solutions that don't involve heavy rewriting and partial/total use of
a skill.

Co-authored-by: Amila Sikalo <askialo2@users.noreply.github.com>